requirements.txt
requests
python-dotenv
pydantic
numpy
rank-bm25
urllib3<2